Rename the mothership component to the reconciler
